added ingredient amount to add dishes to amount

+++ src/models/dishingredients.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   var dishIngredients = sequelize.define('dishIngredients', {
     dishId: DataTypes.INTEGER,
     ingredientId: DataTypes.INTEGER,
-    amount: DataTypes.INTEGER,
+    amount: DataTypes.DOUBLE,
     createdAt: {
       type: DataTypes.DATE,
       defaultValue: DataTypes.NOW
